backend/pyproject.toml
[project]
name = "docify-backend"
version = "0.1.0"
description = "Document intelligence backend — GraphRAG + hybrid Q&A"
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.20",

    # Task queue
    "celery[redis]>=5.4.0",

    # Document parsing
    "pymupdf4llm>=0.0.17",
    "python-docx>=1.1.0",

    # LlamaIndex (vector + embedding abstractions, Phase 2: graph)
    "llama-index-core>=0.12.0",
    "llama-index-vector-stores-qdrant>=0.4.0",
    "llama-index-embeddings-openai>=0.3.0",
    "llama-index-embeddings-ollama>=0.5.0",
    "llama-index-llms-anthropic>=0.6.0",
    "llama-index-llms-ollama>=0.5.0",

    # Vector DB
    "qdrant-client>=1.12.0",

    # LLM
    "anthropic>=0.40.0",

    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.14.0",

    # Config
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
